Moved panic-related queries to lowering/panic module
